Add: parallel generation for towers, Update: rename tmaze::game::game.rs to app.rs
